pyproject.toml
[project]
name = "vortex-ease"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "django>=5.2.9",
    "django-unfold>=0.72.0",
    "gunicorn>=25.1.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "reportlab>=4.0.0",
    "requests>=2.32.5",
    "weasyprint>=61.2",
]
